requirements.txt
Django==5.0.7
reportlab==4.2.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2